Added all 3 post routes for adding data to each table

+++ migrations/20190104191106_lyft_info.js
@@ -3,7 +3,7 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('lyft_info', table => {
     table.increments('id').primary()
     table.integer('eta_of_pickup')
-    table.integer('estimated_price')
+    table.decimal('estimated_price')
   })
   
 };
