+++ migrations/20190205133648_uber_info.js
@@ -2,7 +2,7 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('uber_info', table => {
     table.increments('id').primary()
     table.integer('eta_of_pickup')
-    table.integer('estimated_price')
+    table.decimal('estimated_price')
   })
   
 };
